+++ src/commands/clean-module.js
@@ -2,5 +2,5 @@ import exec from '../exec'
 
 export default function(args) {
   console.log('nwb: clean-module')
-  exec('rimraf', ['es6', 'lib'])
+  exec('rimraf', ['coverage', 'es6', 'lib'])
 }